2019 day 08 and 2019 day 09

+++ src/main/java/uk/co/scottdennison/java/soft/challenges/adventofcode/puzzles/year2019/Day02.java
@@ -13,12 +13,12 @@ import java.util.Arrays;
 public class Day02 implements IPuzzle {
     @Override
     public IPuzzleResults runPuzzle(char[] inputCharacters, IPuzzleConfigProvider configProvider, boolean partBPotentiallyUnsolvable, PrintWriter printWriter) {
-        int[] memory = IntcodeComputer.readProgram(inputCharacters);
-        int part1Result = runProgram(memory,12,2);
-        Integer part2Result = null;
-        for (int noun=0; noun<=99; noun++) {
-            for (int verb=0; verb<=99; verb++) {
-                if (runProgram(memory,noun,verb) == 19690720) {
+        long[] memory = IntcodeComputer.readProgram(inputCharacters);
+        long part1Result = runProgram(memory,12,2);
+        Long part2Result = null;
+        for (long noun=0; noun<=99; noun++) {
+            for (long verb=0; verb<=99; verb++) {
+                if (runProgram(memory,noun,verb) == 19690720L) {
                     if (part2Result != null) {
                         throw new IllegalStateException("Multiple results");
                     }
@@ -32,8 +32,8 @@ public class Day02 implements IPuzzle {
         );
     }
 
-    private static int runProgram(int[] initialMemory, int noun, int verb) {
-        int[] memory = Arrays.copyOf(initialMemory, initialMemory.length);
+    private static long runProgram(long[] initialMemory, long noun, long verb) {
+        long[] memory = Arrays.copyOf(initialMemory, initialMemory.length);
         memory[1] = noun;
         memory[2] = verb;
         IntcodeComputer intcodeComputer = new IntcodeComputer(memory);
